+++ example/android/app/src/main/kotlin/ge/fman/example/MainActivity.kt
@@ -2,5 +2,4 @@ package ge.fman.example
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity : FlutterActivity()