+++ example/lib/main_production.dart
@@ -1,0 +1,6 @@
+import 'package:example/app/app.dart';
+import 'package:example/bootstrap.dart';
+
+Future<void> main() async {
+  await bootstrap(() => const App());
+}